Maybe fixed autonomous, no more duplication- hopefully :+1:, removed debugging op modes from op mode selection, changed things in driverOp, kinda.
