+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/FtcOpModeRegister.java
@@ -35,8 +35,6 @@ package com.qualcomm.ftcrobotcontroller.opmodes;
 import com.qualcomm.ftcrobotcontroller.opmodes.customops.Autonomous.AutonomousMountainRight;
 import com.qualcomm.ftcrobotcontroller.opmodes.customops.Autonomous.AutonomousMountainLeft;
 import com.qualcomm.ftcrobotcontroller.opmodes.customops.TeleOp.DriverOp;
-import com.qualcomm.ftcrobotcontroller.opmodes.customops.MotorTest.MotorTest;
-import com.qualcomm.ftcrobotcontroller.opmodes.customops.MotorTest.ServoTest;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegister;
@@ -63,9 +61,7 @@ public class FtcOpModeRegister implements OpModeRegister {
      */
 
     manager.register("DriverOp", DriverOp.class);
-    manager.register("Autonomous Mountain Left", AutonomousMountainLeft.class);
-    manager.register("Autonomous Mountain Right", AutonomousMountainRight.class);
-    manager.register("[DEBUG ONLY] Dangerous Motor Test [DEBUG ONLY]", MotorTest.class);
-    manager.register("[DEBUG ONLY] Dangerous Servo Test [DEBUG ONLY]", ServoTest.class);
+    manager.register("<- Autonomous Mountain Left", AutonomousMountainLeft.class);
+    manager.register("-> Autonomous Mountain Right", AutonomousMountainRight.class);
   }
 }
