+++ FtcRobotController/src/main/res/layout/activity_ftc_controller.xml
@@ -45,7 +45,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
     android:id="@+id/top_bar"
     android:layout_width="fill_parent"
     android:layout_height="80dp"
-    android:background="@color/pink">
+    android:background="#7b0b0b">
     <ImageView
       android:id="@+id/robotIcon"
       android:layout_width="40dp"
@@ -73,7 +73,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
       android:src="@drawable/icon_menu"
-      android:background="@color/pink"
+      android:background="#b01a1a"
       android:padding="15dp"
       android:adjustViewBounds="true"
       android:layout_alignParentRight="true"
